+++ packages/node-core/src/configure/configure.module.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {handleCreateSubqueryProjectError, LocalReader, makeTempDir, ReaderFactory} from '@subql/common';
-import {Reader} from '@subql/types-core';
+import {IEndpointConfig, Reader} from '@subql/types-core';
 import {camelCase, isNil, omitBy} from 'lodash';
 import {ISubqueryProject} from '../indexer';
 import {getLogger, setDebugFilter} from '../logger';
@@ -38,7 +38,14 @@ export function validDbSchemaName(name: string): boolean {
 // TODO once yargs is in node core we can update
 type Args = Record<string, any>; // typeof yargsOptions.argv['argv']
 
-function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> = {}): Partial<IConfig> {
+function processEndpointConfig(raw?: string | string[]): IEndpointConfig[] {
+  if (!raw) return [];
+  if (typeof raw === 'string') return [JSON.parse(raw)];
+  if (Array.isArray(raw)) return raw.map((raw) => JSON.parse(raw));
+  throw new Error(`Unknown raw value, received: ${raw}`);
+}
+
+export function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> = {}): Partial<IConfig> {
   return Object.entries(yargs).reduce((acc, [key, value]) => {
     if (['_', '$0'].includes(key)) return acc;
 
@@ -49,6 +56,29 @@ function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> = {}): 
         throw new Error('Argument `network-registry` is not valid JSON');
       }
     }
+
+    // Merge network endpoints and possible endpoint configs
+    if (key === 'network-endpoint') {
+      const endpointConfig = processEndpointConfig(yargs['network-endpoint-config']);
+      if (typeof value === 'string') {
+        value = [value];
+      }
+      if (Array.isArray(value)) {
+        value = value.reduce(
+          (acc, endpoint, index) => {
+            acc[endpoint] = endpointConfig[index] ?? {};
+            return acc;
+          },
+          {} as Record<string, IEndpointConfig>
+        );
+      }
+    }
+    if (key === 'primary-network-endpoint') {
+      const endpointConfig = processEndpointConfig(yargs['primary-network-endpoint-config']);
+      value = [value, endpointConfig[0] ?? {}];
+    }
+    if (['network-endpoint-config', 'primary-network-endpoint-config'].includes(key)) return acc;
+
     acc[nameMapping[key] ?? camelCase(key)] = value;
     return acc;
   }, {} as any);
@@ -129,18 +159,23 @@ export async function registerApp<P extends ISubqueryProject>(
     setDebugFilter(config.debug);
   }
 
+  const makeNetworkOverrides = (project?: P) => {
+    // Apply the network endpoint and dictionary from the source project to the parent projects if they are not defined in the config
+    return omitBy(
+      {
+        endpoint: config.networkEndpoints ?? project?.network?.endpoint,
+        dictionary: config.networkDictionaries ?? project?.network?.dictionary,
+      },
+      isNil
+    );
+  };
+
   const project = await createProject(
     config.subquery,
     rawManifest,
     reader,
     await getCachedRoot(reader, config.root),
-    omitBy(
-      {
-        endpoint: config.networkEndpoints,
-        dictionary: config.networkDictionaries,
-      },
-      isNil
-    )
+    makeNetworkOverrides()
   ).catch((err: any) => {
     handleCreateSubqueryProjectError(err, pjson, rawManifest, logger);
     exitWithError(err, logger, 1);
@@ -156,14 +191,7 @@ export async function registerApp<P extends ISubqueryProject>(
       await reader.getProjectSchema(),
       reader,
       await getCachedRoot(reader, config.root),
-      omitBy(
-        // Apply the network endpoint and dictionary from the source project to the parent projects if they are not defined in the config
-        {
-          endpoint: config.networkEndpoints ?? project.network.endpoint,
-          dictionary: config.networkDictionaries ?? project.network.dictionary,
-        },
-        isNil
-      )
+      makeNetworkOverrides(project)
     );
   };