+++ packages/common/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New type for endpoint to allow specifying endpoint options (#2511)
 
 ## [4.1.1] - 2024-07-25
 ### Changed